Move home page into separate component
